requirements.txt
requests
pandas 
sqlalchemy
python-dotenv
fastapi
uvicorn
pathlib
pyarrow
psycopg2-binary